Support for custom headers and for a method's base
Some JSON-RPC servers request more headers, for instance a X-Authentication which is required for each request. Also, the field `method` is sometimes preprended with a base i.e. `ThisLook/LikeAPath/v1.0/methodName`. This changes will allow to specify the method base which will be the same for all requests made with same instantiation of the JsonRPC object.